Updated documentation code example
Update the documentation comment to provide a more meaningful code
example showing the results of a call to parse_link_header().

Also add the relevant data for RFC8288, as it may matter to those who
are trying to determine if the "rel" Relation Type will always be
present in the data they are trying to parse (or not).