Tighten query and param extraction
Tighten up the URI query and header parameter extraction code.  Indexing
into a Vec<> with constant integers runs the risk of causing an
out-of-bounds panic.  Now a malformed query or parameter returns an
Error.

This code will also properly parse a query (or parameter) of the form
foo=bar=baz, which will yield a key of 'foo' and a value of 'bar=baz'.
The previous version would lose the 'baz' portion.