Refactor `parse()`, `parse_with_rel()` for speed
`parse()` and `parse_with_rel()` are very, very similar. The only functional
difference is the key type for the resulting HashMap, but `parse_with_rel()` is
currently implemented as a second-pass operation on top of `parse()` that
wastes time.

This commit takes all the common code between the two functions and separates
it out into a private function `parse_with()` that accepts a closure `make_key`
to translate the key into whatever the desired type is (specificied with
generics).  Then, `parse()` and `parse_with_rel()` just call `parse_with()`
with the appropriate closure.

Also add a test case to catch when `parse_with_rel()` is called on a `Link:`
header that doesn't have a `rel` paramter.