Restore `parse_with_rel_should_err()` test case
Test case `parse_with_rel_should_err()` was made conditional on `#[cfg(not(feature = "url"))]` as the test used a relative ref which is unsupported by feature `url`.  However, the point of the test is to check handling of `Link:` headers which are missing a `rel` parameter, independent of the type of URI being processed.  Thus, we can remove the conditional compilation by fixing the test header to be an absolute URI.