Remove un-needed explicit typing
In Rust, it is generally better to let the compiler infer as many types
as possible.  It makes the code more resiliant to changes in the future.

Functionally, this change makes no difference.  The compiled binary
output is still the same, as the compiler (correctly) infers the same
type that was previously explicitly specified.